requirements.txt
phidata>=2.0.0
groq
python-dotenv
